@@ -1,44 +1,44 @@
 module server
 
-go 1.23.0
-
-toolchain go1.23.5
+go 1.24.0
 
 replace (
-	github.com/anacrolix/torrent v1.58.1 => github.com/tsynik/torrent v1.2.22
+	github.com/anacrolix/torrent v1.59.1 => github.com/tsynik/torrent v1.2.22
 	github.com/anacrolix/upnp v0.1.4 => github.com/tsynik/upnp v0.1.5
 )
 
 require (
 	github.com/agnivade/levenshtein v1.2.1
-	github.com/alexflint/go-arg v1.5.0
-	github.com/anacrolix/dms v1.7.1
-	github.com/anacrolix/log v0.16.0
-	github.com/anacrolix/missinggo/v2 v2.8.0
+	github.com/alexflint/go-arg v1.6.0
+	github.com/anacrolix/dms v1.7.2
+	github.com/anacrolix/log v0.17.0
+	github.com/anacrolix/missinggo/v2 v2.10.0
 	github.com/anacrolix/publicip v0.3.1
-	github.com/anacrolix/torrent v1.58.1
+	github.com/anacrolix/torrent v1.59.1
 	github.com/dustin/go-humanize v1.0.1
-	github.com/gin-contrib/cors v1.7.5
+	github.com/gin-contrib/cors v1.7.6
 	github.com/gin-contrib/location v1.0.3
-	github.com/gin-gonic/gin v1.10.1
+	github.com/gin-gonic/gin v1.11.0
+	github.com/hanwen/go-fuse/v2 v2.9.0
 	github.com/kljensen/snowball v0.10.0
 	github.com/pkg/browser v0.0.0-20240102092130-5ac0b6a4141c
 	github.com/pkg/errors v0.9.1
 	github.com/swaggo/files v1.0.1
-	github.com/swaggo/gin-swagger v1.6.0
-	github.com/swaggo/swag v1.16.4
+	github.com/swaggo/gin-swagger v1.6.1
+	github.com/swaggo/swag v1.16.6
 	github.com/wlynxg/anet v0.0.5
-	go.etcd.io/bbolt v1.4.0
-	golang.org/x/exp v0.0.0-20250606033433-dcc06ee1d476
-	golang.org/x/image v0.28.0
-	golang.org/x/time v0.12.0
-	gopkg.in/telebot.v4 v4.0.0-beta.5
+	go.etcd.io/bbolt v1.4.3
+	golang.org/x/exp v0.0.0-20251125195548-87e1e737ad39
+	golang.org/x/image v0.33.0
+	golang.org/x/net v0.47.0
+	golang.org/x/time v0.14.0
+	gopkg.in/telebot.v4 v4.0.0-beta.7
 	gopkg.in/vansante/go-ffprobe.v2 v2.2.1
 )
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
-	google.golang.org/protobuf v1.36.6 // indirect
+	google.golang.org/protobuf v1.36.10 // indirect
 )
 
 require (
