@@ -25,6 +25,13 @@ export default function AboutDialog() {
 
   const onClose = () => setOpen(false)
   const ref = useOnStandaloneAppOutsideClick(onClose)
+  const getBasePath = () => {
+    if (typeof window !== 'undefined') {
+      return window.location.pathname.split('/')[1] || ''
+    }
+    return ''
+  }
+  const basePath = getBasePath()
 
   return (
     <>
@@ -57,7 +64,7 @@ export default function AboutDialog() {
           <HeaderSection>
             <div>{t('About')}</div>
             {torrServerVersion}
-            <img src='icon.png' alt='ts-icon' />
+            <img src={`${basePath}/icon.png`} alt='ts-icon' />
           </HeaderSection>
 
           <div style={{ overflow: 'auto' }}>
@@ -86,6 +93,7 @@ export default function AboutDialog() {
                 <LinkComponent name='SpAwN_LMG' link='https://github.com/spawnlmg' />
                 <LinkComponent name='damiva' link='https://github.com/damiva' />
                 <LinkComponent name='Vladlenas' link='https://github.com/vladlenas' />
+                <LinkComponent name='Pavel Pikta' link='https://github.com/pavelpikta' />
                 <LinkComponent name='Anton Potekhin' link='https://github.com/Anton111111' />
                 <LinkComponent name='FaintGhost' link='https://github.com/FaintGhost' />
                 <LinkComponent name='TopperBG' link='https://github.com/TopperBG' />
