@@ -23,8 +23,9 @@ export default {
   UploadRateLimit: 0,
   ConnectionsLimit: 25,
   PeersListenPort: 0,
-  ResponsiveMode: false,
+  ResponsiveMode: true,
   SslPort: 0,
   SslCert: '',
   SslKey: '',
+  ShowFSActiveTorr: true,
 }
